+++ packages/components/src/Popover/Layout/PopoverContent/index.ts
@@ -24,38 +24,4 @@
 
  */
 
-import React, { FC } from 'react'
-import { ArrowDropDown } from '@styled-icons/material/ArrowDropDown'
-import { ButtonOutline } from '../../Button'
-import { Box } from '../../Layout'
-import { Popover } from '../Popover'
-import { PopoverContent } from '../Layout/PopoverContent/PopoverContent'
-
-interface Props {
-  top?: number
-  left?: number
-  bottom?: number
-  right?: number
-}
-
-export const EdgeOverflow: FC<Props> = ({
-  children,
-  top,
-  left,
-  bottom,
-  right,
-}) => (
-  <Box position="absolute" top={top} left={left} bottom={bottom} right={right}>
-    <Popover
-      content={
-        <PopoverContent width="18rem" height="5rem">
-          There's stuff here... it hits the edge{' '}
-        </PopoverContent>
-      }
-    >
-      <ButtonOutline iconAfter={<ArrowDropDown />} m="xxlarge">
-        {children}
-      </ButtonOutline>
-    </Popover>
-  </Box>
-)
+export * from './PopoverContent'
